Phase 3 commit 1: extract data/http/* adapters and data/ports.ts
Mirror the server-side layered structure on the client:

- data/ports.ts — type-only port interfaces (EventStore, StateStore,
  NoteStore, LinkStore). Documents the contract a future port-object
  adapter would expose; today's flat-function adapters match.
- data/http/client.ts — ApiError + jsonFetch + rawFetch. Sole
  sanctioned client HTTP path.
- data/http/events.http.ts — listEvents / getEvent / createEvent /
  updateEvent / deleteEvent.
- data/http/state.http.ts — state / tags / palette / sessions
  GET/PUT plus appendSession.
- data/http/links.http.ts — getLinkIndex, getFile (peek).
- data/http/notes.http.ts — folders + files + assets.

data/api.ts collapses to a thin re-export aggregator so existing
callers (10 files) keep working without changes. The next commit
migrates those callers to import from the new locations directly
and deletes api.ts.

No behaviour change. Tests baseline unchanged.
